proj1

+++ proj1/ColorMapInput.h
@@ -10,6 +10,6 @@
 **This function reads in a file name colorfile.
 **It then uses the information in colorfile to create a color array, with each color represented by an int[3].
 ***************/
-uint8_t** FileToColorMap(char* colorfile, int* colorcount);
+int** FileToColorMap(char* colorfile, int* colorcount);
